+++ finalproject/pubspec.yaml
@@ -56,6 +56,9 @@ dev_dependencies:
 
 # The following section is specific to Flutter packages.
 flutter:
+  assets:
+      - assets/google.png
+      - assets/dx logo.png
 
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
